SCRUM -011-CI-Updated the collab test
