+++ front-end/pubspec.yaml
@@ -49,7 +49,10 @@ dependencies:
 
 
 dev_dependencies:
-  mockito: ^5.4.4
+
+  mockito: ^5.4.2
+  build_runner: ^2.4.6
+
   provider: ^6.1.0
   flutter_test:
     sdk: flutter
